Restructure to create post processor class.
