+++ koku/masu/util/aws/common.py
@@ -4,14 +4,12 @@
 #
 """AWS utility functions."""
 import datetime
-import json
 import logging
 import re
 import uuid
 from itertools import chain
 
 import boto3
-import ciso8601
 import pandas as pd
 from botocore.config import Config
 from botocore.exceptions import ClientError
@@ -25,10 +23,7 @@ from api.provider.models import Provider
 from masu.database.aws_report_db_accessor import AWSReportDBAccessor
 from masu.database.provider_db_accessor import ProviderDBAccessor
 from masu.util import common as utils
-from masu.util.common import safe_float
-from masu.util.common import strip_characters_from_column_name
 from masu.util.ocp.common import match_openshift_labels
-from reporting.provider.aws.models import TRINO_REQUIRED_COLUMNS
 
 LOG = logging.getLogger(__name__)
 
@@ -92,6 +87,105 @@ CSV_COLUMN_PREFIX = (
     "pricing/",
     "discount/",
     "product/sku",
+)
+
+INGRESS_REQUIRED_COLUMNS = set(
+    "bill/BillingEntity",
+    "bill/BillType",
+    "bill/PayerAccountId",
+    "bill/BillingPeriodStartDate",
+    "bill/BillingPeriodEndDate",
+    "bill/InvoiceId",
+    "lineItem/LineItemType",
+    "lineItem/UsageAccountId",
+    "lineItem/UsageStartDate",
+    "lineItem/UsageEndDate",
+    "lineItem/ProductCode",
+    "lineItem/UsageType",
+    "lineItem/Operation",
+    "lineItem/AvailabilityZone",
+    "lineItem/ResourceId",
+    "lineItem/UsageAmount",
+    "lineItem/NormalizationFactor",
+    "lineItem/NormalizedUsageAmount",
+    "lineItem/CurrencyCode",
+    "lineItem/UnblendedRate",
+    "lineItem/UnblendedCost",
+    "lineItem/BlendedRate",
+    "lineItem/BlendedCost",
+    "savingsPlan/SavingsPlanEffectiveCost",
+    "lineItem/TaxType",
+    "pricing/publicOnDemandCost",
+    "pricing/publicOnDemandRate",
+    "reservation/AmortizedUpfrontFeeForBillingPeriod",
+    "reservation/AmortizedUpfrontCostForUsage",
+    "reservation/RecurringFeeForUsage",
+    "reservation/UnusedQuantity",
+    "reservation/UnusedRecurringFee",
+    "pricing/term",
+    "pricing/unit",
+    "product/sku",
+    "product/ProductName",
+    "product/productFamily",
+    "product/servicecode",
+    "product/region",
+    "product/instanceType",
+    "product/memory",
+    "product/vcpu",
+    "reservation/ReservationARN",
+    "reservation/NumberOfReservations",
+    "reservation/UnitsPerReservation",
+    "reservation/StartTime",
+    "reservation/EndTime",
+)
+
+INGRESS_ALT_COLUMNS = set(
+    "bill_billing_entity",
+    "bill_bill_type",
+    "bill_payer_account_id",
+    "bill_billing_period_start_date",
+    "bill_billing_period_end_date",
+    "bill_invoice_id",
+    "line_item_line_item_type",
+    "line_item_usage_account_id",
+    "line_item_usage_start_date",
+    "line_item_usage_end_date",
+    "line_item_product_code",
+    "line_item_usage_type",
+    "line_item_operation",
+    "line_item_availability_zone",
+    "line_item_resource_id",
+    "line_item_usage_amount",
+    "line_item_normalization_factor",
+    "line_item_normalized_usage_amount",
+    "line_item_currency_code",
+    "line_item_unblended_rate",
+    "line_item_unblended_cost",
+    "line_item_blended_rate",
+    "line_item_blended_cost",
+    "savings_plan_savings_plan_effective_cost",
+    "line_item_tax_type",
+    "pricing_public_on_demand_cost",
+    "pricing_public_on_demand_rate",
+    "reservation_amortized_upfront_fee_for_billing_period",
+    "reservation_amortized_upfront_cost_for_usage",
+    "reservation_recurring_fee_for_usage",
+    "reservation_unused_quantity",
+    "reservation_unused_recurring_fee",
+    "pricing_term",
+    "pricing_unit",
+    "product_sku",
+    "product_product_name",
+    "product_product_family",
+    "product_servicecode",
+    "product_region",
+    "product_instance_type",
+    "product_memory",
+    "product_vcpu",
+    "reservation_number_of_reservations",
+    "reservation_units_per_reservation",
+    "reservation_start_time",
+    "reservation_end_time",
 )
 
 
@@ -549,115 +643,6 @@ def remove_files_not_in_set_from_s3_bucket(request_id, s3_path, manifest_id, con
     return removed
 
 
-def handle_user_defined_json_columns(data_frame, columns, column_prefix):
-    """Given a prefix convert multiple dataframe columns into a single json column."""
-
-    def scrub_resource_col_name(res_col_name):
-        return res_col_name.replace(column_prefix, "")
-
-    columns_of_interest = [column for column in columns if column_prefix in column]
-    unique_keys = {scrub_resource_col_name(column) for column in columns_of_interest}
-
-    df = data_frame[columns_of_interest]
-    column_dict = df.apply(
-        lambda row: {scrub_resource_col_name(column): value for column, value in row.items() if value}, axis=1
-    )
-    column_dict.where(column_dict.notna(), lambda _: [{}], inplace=True)
-
-    return column_dict.apply(json.dumps), unique_keys
-
-
-def check_aws_custom_columns(col_names):
-    REQUIRED_COLS = set(utils.CSV_ALT_COLUMNS["AWS"])
-    missing_cols = False
-    if not set(col_names).issuperset(REQUIRED_COLS):
-        missing_cols = True
-    return missing_cols, REQUIRED_COLS
-
-
-def aws_post_processor(data_frame):
-    """
-    Consume the AWS data and add a column creating a dictionary for the aws tags
-    """
-    org_columns = data_frame.columns.unique()
-    columns = []
-    for col in org_columns:
-        if "/" not in col and COL_TRANSLATION.get(col):
-            data_frame = data_frame.rename(columns={col: COL_TRANSLATION[col]})
-            columns.append(COL_TRANSLATION[col])
-    columns = set(TRINO_REQUIRED_COLUMNS).union(data_frame)
-    columns = sorted(list(columns))
-
-    tags, unique_keys = handle_user_defined_json_columns(data_frame, columns, RESOURCE_TAG_USER_PREFIX)
-    data_frame["resourceTags"] = tags
-
-    cost_categories, aws_category_keys = handle_user_defined_json_columns(data_frame, columns, COST_CATEGORY_PREFIX)
-    data_frame["costCategory"] = cost_categories
-
-    # Make sure we have entries for our required columns
-    data_frame = data_frame.reindex(columns=columns)
-
-    columns = list(data_frame)
-    column_name_map = {}
-    drop_columns = []
-    for column in columns:
-        new_col_name = strip_characters_from_column_name(column)
-        column_name_map[column] = new_col_name
-        if ALL_RESOURCE_TAG_PREFIX in column or COST_CATEGORY_PREFIX in column:
-            drop_columns.append(column)
-    data_frame = data_frame.drop(columns=drop_columns)
-    data_frame = data_frame.rename(columns=column_name_map)
-    return (data_frame, unique_keys, aws_category_keys)
-
-
-def aws_generate_daily_data(data_frame):
-    """Given a dataframe, group the data to create daily data."""
-    # usage_start = data_frame["lineitem_usagestartdate"]
-    # usage_start_dates = usage_start.apply(lambda row: row.date())
-    # data_frame["usage_start"] = usage_start_dates
-    daily_data_frame = data_frame.groupby(
-        [
-            "lineitem_resourceid",
-            pd.Grouper(key="lineitem_usagestartdate", freq="D"),
-            "bill_payeraccountid",
-            "lineitem_usageaccountid",
-            "lineitem_legalentity",
-            "lineitem_lineitemdescription",
-            "bill_billingentity",
-            "lineitem_productcode",
-            "lineitem_availabilityzone",
-            "product_productfamily",
-            "product_instancetype",
-            "product_region",
-            "pricing_unit",
-            "resourcetags",
-            "costcategory",
-        ],
-        dropna=False,
-    ).agg(
-        {
-            "lineitem_usageamount": ["sum"],
-            "lineitem_normalizationfactor": ["max"],
-            "lineitem_normalizedusageamount": ["sum"],
-            "lineitem_currencycode": ["max"],
-            "lineitem_unblendedrate": ["max"],
-            "lineitem_unblendedcost": ["sum"],
-            "lineitem_blendedrate": ["max"],
-            "lineitem_blendedcost": ["sum"],
-            "pricing_publicondemandcost": ["sum"],
-            "pricing_publicondemandrate": ["max"],
-            "savingsplan_savingsplaneffectivecost": ["sum"],
-            "product_productname": ["max"],
-            "bill_invoiceid": ["max"],
-        }
-    )
-    columns = daily_data_frame.columns.droplevel(1)
-    daily_data_frame.columns = columns
-    daily_data_frame.reset_index(inplace=True)
-
-    return daily_data_frame
-
-
 def match_openshift_resources_and_labels(data_frame, cluster_topologies, matched_tags):
     """Filter a dataframe to the subset that matches an OpenShift source."""
     resource_ids = chain.from_iterable(
@@ -716,37 +701,3 @@ def match_openshift_resources_and_labels(data_frame, cluster_topologies, matched
     )
 
     return openshift_matched_data_frame
-
-
-def get_column_converters():
-    """Return source specific parquet column converters."""
-    return {
-        "bill/billingperiodstartdate": ciso8601.parse_datetime,
-        "bill/billingperiodenddate": ciso8601.parse_datetime,
-        "lineitem/usagestartdate": ciso8601.parse_datetime,
-        "lineitem/usageenddate": ciso8601.parse_datetime,
-        "lineitem/usageamount": safe_float,
-        "lineitem/normalizationfactor": safe_float,
-        "lineitem/normalizedusageamount": safe_float,
-        "lineitem/unblendedrate": safe_float,
-        "lineitem/unblendedcost": safe_float,
-        "lineitem/blendedrate": safe_float,
-        "lineitem/blendedcost": safe_float,
-        "pricing/publicondemandcost": safe_float,
-        "pricing/publicondemandrate": safe_float,
-        "savingsplan/savingsplaneffectivecost": safe_float,
-        "bill_billing_period_start_date": ciso8601.parse_datetime,
-        "bill_billing_period_end_date": ciso8601.parse_datetime,
-        "line_item_usage_start_date": ciso8601.parse_datetime,
-        "line_item_usage_end_date": ciso8601.parse_datetime,
-        "line_item_usage_amount": safe_float,
-        "line_item_normalization_factor": safe_float,
-        "line_item_normalized_usage_amount": safe_float,
-        "line_item_unblended_rate": safe_float,
-        "line_item_unblended_cost": safe_float,
-        "line_item_blended_rate": safe_float,
-        "line_item_blended_cost": safe_float,
-        "pricing_public_on_demand_cost": safe_float,
-        "pricing_public_on_demand_rate": safe_float,
-        "savings_plan_savings_plan_effective_cost": safe_float,
-    }
