+++ koku/masu/util/gcp/common.py
@@ -4,12 +4,9 @@
 #
 """GCP utility functions and vars."""
 import datetime
-import json
 import logging
 import uuid
-from json.decoder import JSONDecodeError
 
-import ciso8601
 import pandas as pd
 from tenant_schemas.utils import schema_context
 
@@ -17,13 +14,43 @@ from api.models import Provider
 from masu.database.provider_db_accessor import ProviderDBAccessor
 from masu.external.accounts_accessor import AccountsAccessor
 from masu.processor import disable_gcp_resource_matching
-from masu.util.common import safe_float
-from masu.util.common import strip_characters_from_column_name
 from masu.util.ocp.common import match_openshift_labels
 from reporting.provider.gcp.models import GCPCostEntryBill
 
 LOG = logging.getLogger(__name__)
 pd.options.mode.chained_assignment = None
+
+INGRESS_REQUIRED_COLUMNS = set(
+    "billing_account_id",
+    "service.id",
+    "service.description",
+    "sku.id",
+    "sku.description",
+    "usage_start_time",
+    "usage_end_time",
+    "project.id",
+    "project.name",
+    "project.labels",
+    "project.ancestry_numbers",
+    "labels",
+    "system_labels",
+    "location.location",
+    "location.country",
+    "location.region",
+    "location.zone",
+    "export_time",
+    "cost",
+    "currency",
+    "currency_conversion_rate",
+    "usage.amount",
+    "usage.unit",
+    "usage.amount_in_pricing_units",
+    "usage.pricing_unit",
+    "credits",
+    "invoice.month",
+    "cost_type",
+    "partition_date",
+)
 
 
 def get_bills_from_provider(provider_uuid, schema, start_date=None, end_date=None):
@@ -69,129 +96,6 @@ def get_bills_from_provider(provider_uuid, schema, start_date=None, end_date=Non
         bills = bills.all()
 
     return bills
-
-
-def process_gcp_labels(label_string):
-    """Convert the report string to a JSON dictionary.
-
-    Args:
-        label_string (str): The raw report string of pod labels
-
-    Returns:
-        (dict): The JSON dictionary made from the label string
-
-    """
-    label_dict = {}
-    try:
-        if label_string:
-            labels = json.loads(label_string)
-            label_dict = {entry.get("key"): entry.get("value") for entry in labels}
-    except JSONDecodeError:
-        LOG.warning("Unable to process GCP labels.")
-
-    return json.dumps(label_dict)
-
-
-def process_gcp_credits(credit_string):
-    """Process the credits column, which is non-standard JSON."""
-    credit_dict = {}
-    try:
-        credits = json.loads(credit_string.replace("'", '"').replace("None", '"None"'))
-        if credits:
-            credit_dict = credits[0]
-    except JSONDecodeError:
-        LOG.warning("Unable to process GCP credits.")
-
-    return json.dumps(credit_dict)
-
-
-def gcp_post_processor(data_frame):
-    """Guarantee column order for GCP parquet files"""
-    columns = list(data_frame)
-    column_name_map = {}
-    for column in columns:
-        new_col_name = strip_characters_from_column_name(column)
-        column_name_map[column] = new_col_name
-    data_frame = data_frame.rename(columns=column_name_map)
-
-    label_set = set()
-    unique_labels = data_frame.labels.unique()
-    for label in unique_labels:
-        label_set.update(json.loads(label).keys())
-
-    return (data_frame, label_set, None)
-
-
-def get_column_converters():
-    """Return source specific parquet column converters."""
-    return {
-        "usage_start_time": ciso8601.parse_datetime,
-        "usage_end_time": ciso8601.parse_datetime,
-        "project.labels": process_gcp_labels,
-        "labels": process_gcp_labels,
-        "system_labels": process_gcp_labels,
-        "export_time": ciso8601.parse_datetime,
-        "cost": safe_float,
-        "currency_conversion_rate": safe_float,
-        "usage.amount": safe_float,
-        "usage.amount_in_pricing_units": safe_float,
-        "credits": process_gcp_credits,
-    }
-
-
-def gcp_generate_daily_data(data_frame):
-    """Given a dataframe, return the data frame if its empty, group the data to create daily data."""
-    if data_frame.empty:
-        return data_frame
-
-    # this parses the credits column into just the dollar amount so we can sum it up for daily rollups
-    rollup_frame = data_frame.copy()
-    rollup_frame["credits"] = rollup_frame["credits"].apply(json.loads)
-    rollup_frame["daily_credits"] = 0.0
-    for i, credit_dict in enumerate(rollup_frame["credits"]):
-        rollup_frame["daily_credits"][i] = credit_dict.get("amount", 0.0)
-    resource_df = rollup_frame.get("resource_name")
-    try:
-        if not resource_df:
-            rollup_frame["resource_name"] = ""
-            rollup_frame["resource_global_name"] = ""
-    except Exception:
-        if not resource_df.any():
-            rollup_frame["resource_name"] = ""
-            rollup_frame["resource_global_name"] = ""
-    daily_data_frame = rollup_frame.groupby(
-        [
-            "invoice_month",
-            "billing_account_id",
-            "project_id",
-            pd.Grouper(key="usage_start_time", freq="D"),
-            "service_id",
-            "sku_id",
-            "system_labels",
-            "labels",
-            "cost_type",
-            "location_region",
-            "resource_name",
-        ],
-        dropna=False,
-    ).agg(
-        {
-            "project_name": ["max"],
-            "service_description": ["max"],
-            "sku_description": ["max"],
-            "usage_pricing_unit": ["max"],
-            "usage_amount_in_pricing_units": ["sum"],
-            "currency": ["max"],
-            "cost": ["sum"],
-            "daily_credits": ["sum"],
-            "resource_global_name": ["max"],
-        }
-    )
-    columns = daily_data_frame.columns.droplevel(1)
-    daily_data_frame.columns = columns
-    daily_data_frame.reset_index(inplace=True)
-
-    return daily_data_frame
 
 
 def match_openshift_resources_and_labels(data_frame, cluster_topologies, matched_tags):
